feat: add cjsToEsm task
