+++ package.json
@@ -11,7 +11,8 @@
     "fs:delete": "node src/fs/delete.js",
     "fs:list": "node src/fs/list.js",
     "fs:read": "node src/fs/read.js",
-    "cli:args": "node src/cli/args.js --propName value --prop2Name value2"
+    "cli:args": "node src/cli/args.js --propName value --prop2Name value2",
+    "modules:cjsToEsm": "node src/modules/cjsToEsm.mjs"
   },
   "repository": {
     "type": "git",
